Add "NPC alarm" feature to Town Hall
If a redstone signal is sent to the Town Hall (indirectly, i.e. vanilla style) then all NPC's within the broadcast range of that town center will be ordered to return to their respective home. Redstone signal disabled will broadcast an "all clear".

A highly wanted feature of myself and @Dulciphi so third-party mods can declare a "disaster mode", for example the Tornado Sensor in Wheather2 mod and DecoCraft's Fire Alarm for example.