Make Upkeep Order UX a bit friendlier
Will give a chat message if the upkeep point is set successfully, else open the GUI (instead of just opening the GUI every time, which incorrectly shows the old upkeep point in the event of it being changed).

This is also more in-line with the UX of other order slips.