Add "Headquarters" block [...]
Requires FTBU_AW2 - HQ will not exist if it's not loaded.

The first Town Hall placed by a player becomes their Headquarters. If the HQ is lost/broken, then the next placed Town Hall becomes they HQ.

For the moment, the only difference between HQ and Town Hall is that HQ is blue. Reign-modpack specific stuff is planned.